[CELEBORN-1190][FOLLOWUP] Apply error prone patch and suppress some problems
1. Fix IntLongMath, InconsistentCapitalization, UnnecessaryAssignment
2. disable StringSplitter, EmptyBlockTag, EqualsGetClass, MissingSummary, BadImport

No

GA

```bash
./build/make-distribution.sh --release
```

PR

total  202

```
  43 SynchronizeOnNonFinalField
  24 StaticAssignmentInConstructor
  21 JdkObsolete
  16 ThreadLocalUsage
  16 MutableConstantField
  16 MissingCasesInEnumSwitch
  11 UnusedMethod
  11 NonAtomicVolatileUpdate
   8 UnusedNestedClass
   8 NonOverridingEquals
   8 Finally
   4 MixedMutabilityReturnType
   4 DoubleBraceInitialization
   4 CatchAndPrintStackTrace
   4 CanonicalDuration
   2 ReferenceEquality
   1 ClassCanBeStatic
   1 ByteBufferBackingArray
```

Closes #2180 from cxzl25/error_prone_patch_followup.

Authored-by: sychen <sychen@ctrip.com>

+++ common/src/main/java/org/apache/celeborn/common/network/util/TransportFrameDecoder.java
@@ -97,7 +97,7 @@ public class TransportFrameDecoder extends ChannelInboundHandlerAdapter implemen
       msgSize = first.readInt();
       curType = Message.Type.decode(first);
       bodySize = first.readInt();
-      nextFrameSize = msgSize + bodySize;
+      nextFrameSize = (long) msgSize + bodySize;
       totalSize -= HEADER_SIZE;
       if (!first.isReadable()) {
         buffers.removeFirst().release();
@@ -117,7 +117,7 @@ public class TransportFrameDecoder extends ChannelInboundHandlerAdapter implemen
     msgSize = headerBuf.readInt();
     curType = Message.Type.decode(headerBuf);
     bodySize = headerBuf.readInt();
-    nextFrameSize = msgSize + bodySize;
+    nextFrameSize = (long) msgSize + bodySize;
     totalSize -= HEADER_SIZE;
     return nextFrameSize;
   }
